fix bug in selfification of submodules with a "where type" module type.
If we had
  module X :: {
    type T
    module Y :: SIG where type p = T
  }

We did not add X.T to the global environment while trying to whnf
  "SIG where type p = X.T" and got an error message about X.T not being
  defined.

Solution is to extend the environment while selfifying (and also while
traversing a signature digging for the path to patch).

Unfortunately while keeping the code modular this introduces a circular
dependency between WhnfModuleType, ExtendModuleCtx and Selfify modules.
Solution is to add a boot file for ExtendModuleCtx.