Better translation for datatype value constructors.
Represent datatypes as a record of a dataIn and a dataOut field where
the former is a product of constructor functions and the latter is a
coercion from the datatype abstract var to a sum type of constructor
records.

Update Insomnia type inference to annotate global value constructors
with the datatype that they came from.