[minor] Use fresh names when naming intermediate module values in FOmega.
We often end up with things like:
  let M = { ... }
  in let M = M
  in { ... }

This is legal, but when we translate to Gamble it's slighly less work if
  we see:

  let M' = { ...} in let M = M' in { ... }

(at the toplevel in Gamble, I want to translate this into a sequence of
defines, and they are all mutually in scope at the same time, so unique
names are better.  I could take care of it in the Gamble translation,
but I'd rather not bother unless I have to.